requirements.txt
django==4.0.4
djangorestframework==3.13.1
cloudinary==1.29.0
dj3-cloudinary-storage==0.0.6
djangorestframework_simplejwt==5.1.0
dj_database_url==0.5.0
drf-spectacular==0.22.0
gunicorn==20.1.0
pillow==9.1.0
python-dotenv==0.20.0
weasyprint==54.3
whitenoise==6.0.0

